+++ src/lib/index.js
@@ -8,6 +8,9 @@ const { Importer, ImportResult, ERROR_MODE } = require('./importer');
 const { Pipe, PipeResult, generateDDL }      = require('./pipe');
 const { listDrivers, getDriver, DRIVERS }    = require('./drivers');
 const { Dialect }                            = require('./dialect');
+const { runMigrations }                      = require('./migrations/migrationRunner');
+const { runSeeds }                           = require('./migrations/seedRunner');
+const { createMigration, createSeed }        = require('./migrations/create');
 
 module.exports = {
   ODBCConnection,
@@ -33,4 +36,8 @@ module.exports = {
   getDriver,
   DRIVERS,
   Dialect,
+  runMigrations,
+  runSeeds,
+  createMigration,
+  createSeed,
 };